backend/scraper/views.py
import json
import requests
from bs4 import BeautifulSoup
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework import status
from .models import ScrapedProduct
from .serializers import ScrapedProductSerializer

class ProductSearchView(APIView):
    def post(self, request):
        try:
            # Get search parameters
            keyword = request.data.get('keyword', '')
            min_price = float(request.data.get('minPrice', 0))
            max_price = float(request.data.get('maxPrice', float('inf')))
            min_rating = float(request.data.get('minRating', 0))
            is_prime_only = request.data.get('isPrime', False)

            # Perform web scraping (example implementation)
            products = self.scrape_amazon(keyword)

            # Filter results
            filtered_products = [
                product for product in products
                if (min_price <= product['price'] <= max_price and
                    product['rating'] >= min_rating and
                    (not is_prime_only or product['is_prime']))
            ]

            # Save results to file
            self.save_to_file(filtered_products)

            # Save to database and serialize
            saved_products = []
            for product_data in filtered_products:
                product = ScrapedProduct.objects.create(**product_data)
                saved_products.append(product)

            serializer = ScrapedProductSerializer(saved_products, many=True)
            return Response(serializer.data)

        except Exception as e:
            return Response(
                {'error': str(e)},
                status=status.HTTP_500_INTERNAL_SERVER_ERROR
            )

    def scrape_amazon(self, keyword):
        # This is a placeholder implementation
        # In a real application, you would implement proper web scraping logic
        # using requests and BeautifulSoup
        
        headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
        }
        
        # Example scraping logic (not implemented for demonstration)
        products = []
        # Actual scraping code would go here
        return products

    def save_to_file(self, products):
        with open('results.txt', 'w') as f:
            json.dump(products, f, indent=2)