backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
django-cors-headers==4.3.1